remodeled for CallerID recognition
Remodeled the app for the purpose of CallerID recognition with support of Connexant chip based modem. Each call produces a file with data/time and cid in file's name. There is additional file with all numbers and a debug file with all communication data. App has auto hide to tray featue after startup